requirements.txt
# Core
torch
torchvision
numpy
pillow

# Anomaly Detection
anomalib

# Data and Visualization
matplotlib
seaborn
pandas
opencv-python
streamlit

# Metrics
scikit-learn
scikit-image

# Experiment tracking
wandb
tensorboard

# Notebooks
jupyter
ipywidgets

# Utils
pyyaml
tqdm